extend fix to attest flow
